tesseract/requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
pytesseract>=0.3.10
google-cloud-vision>=3.4.4
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
